requirements.txt
Flask==3.0.3
pymongo==4.7.2
mongomock==4.1.2
python-dotenv==1.0.1
Werkzeug==3.0.3
